Allow use of local box for base box.
config.vm.box refers to the box that is used as the starting
point for the box being brought up. That later box gets its
own name and can be packaged.

By getting rid of the vm.box assignment it becomes possible to
use a box that has been added to the local vagrant set up as
the base.

The Vagrantfile has been adjusted so that if conf['box_name']
is set it wins over conf['box_url'] with a fallback to box_url
and then a fallback to the internally defined precise box.

config.yaml.sample comments updated accordingly.

Also took advantage of this moment to moan about Vagrant's
handling of network interface identifiers (in the config sample).