safely handle git clone from alien directory
git clone can experience permission errors if called from a
cwd which is not the parent of the target for the clone.

This is hard to replicate, but seems to have something to
do with the underlying user executing puppet not being 'stack'
(e.g. 'vagrant').

When the user is 'stack' the cwd is probably /home/stack anyway.

In any case, this fixes it.