Add suport for vagrant-hostmanager to manage /etc/hosts
At the moment the used hostnames in the configuration file are
maybe not resolvable on the Vagrant instances which will result
in issues (for example the SERVICE_HOST variable in local.conf
used the hostname of the manager node).

The vagrant-hostmanager plugin manages the /etc/hosts file
and will set correct entries there to make the hostnames in the
configuration file resolvable.

Change-Id: I7ecd379c9f7e47c04abc85aa61de9c6c50f8a50b